+++ include/distortos/internal/scheduler/Scheduler.hpp
@@ -49,9 +49,6 @@ public:
 	 *
 	 * ThreadControlBlock's state is changed to "runnable".
 	 *
-	 * \param [in] run is a reference to Thread's "run" function
-	 * \param [in] preTerminationHook is a pointer to Thread's pre-termination hook, nullptr to skip
-	 * \param [in] terminationHook is a reference to Thread's termination hook
 	 * \param [in] threadControlBlock is a reference to added ThreadControlBlock object
 	 *
 	 * \return 0 on success, error code otherwise:
@@ -60,8 +57,7 @@ public:
 	 * - error codes returned by Stack::initialize();
 	 */
 
-	int add(void (& run)(Thread&), void (* preTerminationHook)(Thread&), void (& terminationHook)(Thread&),
-			ThreadControlBlock& threadControlBlock);
+	int add(ThreadControlBlock& threadControlBlock);
 
 	/**
 	 * \brief Blocks current thread, transferring it to provided container.