+++ include/distortos/internal/scheduler/Stack.hpp
@@ -23,10 +23,10 @@
 namespace distortos
 {
 
-class Thread;
-
 namespace internal
 {
+
+class RunnableThread;
 
 /// Stack class is an abstraction of architecture's stack
 class Stack
@@ -119,17 +119,13 @@ public:
 	/**
 	 * \brief Fills the stack with stack sentinel, initializes its contents and stack pointer value.
 	 *
-	 * \param [in] thread is a reference to Thread object passed to function
-	 * \param [in] run is a reference to Thread's "run" function
-	 * \param [in] preTerminationHook is a pointer to Thread's pre-termination hook, nullptr to skip
-	 * \param [in] terminationHook is a reference to Thread's termination hook
+	 * \param [in] runnableThread is a reference to RunnableThread object that is being run
 	 *
 	 * \return 0 on success, error code otherwise:
 	 * - error codes returned by architecture::initializeStack();
 	 */
 
-	int initialize(Thread& thread, void (& run)(Thread&), void (* preTerminationHook)(Thread&),
-			void (& terminationHook)(Thread&));
+	int initialize(RunnableThread& runnableThread);
 
 	/**
 	 * \brief Sets value of stack pointer.